+++ Sql/Schema/Tables/Articles/Articles.sql
@@ -3,10 +3,10 @@
     (
 	    [articleId] INT NOT NULL PRIMARY KEY, 
         [feedId] INT NULL DEFAULT 0, 
-	    [subjects] TINYINT NULL DEFAULT 0,
+	    [subjects] SMALLINT NULL DEFAULT 0,
 	    [subjectId] INT NULL DEFAULT 0,
 	    [score] SMALLINT NULL DEFAULT 0,
-        [images] TINYINT NULL DEFAULT 0, 
+        [images] SMALLINT NULL DEFAULT 0, 
 	    [filesize] FLOAT NULL DEFAULT 0,
         [linkcount] INT DEFAULT 0, 
         [linkwordcount] INT DEFAULT 0,
